+++ src/Azos/Log/ArchiveConventions.cs
@@ -23,26 +23,6 @@ namespace Azos.Log
   /// </summary>
   public static class ArchiveConventions
   {
-    /// <summary>
-    /// Structured data bix binary version tag. Serializer writes this as the first field of structured data stream.
-    /// Deserializer reads it and can adjust the reading format for backward compatibility
-    /// </summary>
-    public const byte SD_BIX_VERSION = 1;
-
-    /// <summary>
-    /// Absolute limit on maximum number of map properties per level
-    /// </summary>
-    public const int SD_MAX_MAP_PROPS = 512;
-
-    /// <summary>
-    /// Absolute limit on maximum number of array elements
-    /// </summary>
-    public const int SD_MAX_ARRAY_ELM = 4 * 1024;
-
-    // Null/NonNull map flags, also used as a stream alignment crosscheck code
-    private const byte SD_FMT_NULL = 0x55;
-    private const byte SD_FMT_NOTNULL = 0xAA;
-
     private static readonly JsonWritingOptions AD_JSON_ENCODE_FORMAT = new (JsonWritingOptions.CompactRowsAsMap)
     {
       MapSkipNulls = true,
@@ -237,264 +217,5 @@ namespace Azos.Log
 
       return fact;
     }
-
-    /// <summary>
-    /// Reads a map written in a canonical format which is based on Bix wire primitives.
-    /// Null value may be returned if it was written that way
-    /// </summary>
-    public static JsonDataMap ReadArchivedDataMap(BixReader reader)
-    {
-      var ver = reader.ReadByte();
-      return readArchivedDataMap(reader, ver);
-    }
-
-    private static JsonDataMap readArchivedDataMap(BixReader reader, byte ver)
-    {
-      var flagNullNotNull = reader.ReadByte();
-      if (flagNullNotNull == SD_FMT_NULL) return null;
-      Aver.AreEqual(SD_FMT_NOTNULL, flagNullNotNull, "corrupted data");
-
-      var caseSensitive = reader.ReadBool();
-      var result = new JsonDataMap(caseSensitive);
-
-      while(true)
-      {
-        var key = reader.ReadString();
-        if (key == null) break;//eof
-        var val = readValue(reader, ver);
-        Aver.IsTrue(result.Count < SD_MAX_MAP_PROPS, "max props");
-        result[key] = val;
-      }
-
-      return result;
-    }
-
-    private static JsonDataArray readArchivedDataArray(BixReader reader, byte ver)
-    {
-      var flagNullNotNull = reader.ReadByte();
-      if (flagNullNotNull == SD_FMT_NULL) return null;
-      Aver.AreEqual(SD_FMT_NOTNULL, flagNullNotNull, "corrupted data");
-
-      var count = reader.ReadInt();
-      Aver.IsTrue(count < SD_MAX_ARRAY_ELM, "max elm");
-      var result = new JsonDataArray();
-
-      for (var i=0; i<count; i++)
-      {
-        var val = readValue(reader, ver);
-        result.Add(val);
-      }
-
-      return result;
-    }
-
-    /// <summary>
-    /// Writes map into a stream using a canonical archive format which is based on Bix wire primitives.
-    /// Null is permitted. The sub-objects should be wither JsonDataMap or JsonDataArray
-    /// </summary>
-    public static void WriteArchivedDataMap(BixWriter writer, JsonDataMap map)
-    {
-      writer.Write(SD_BIX_VERSION);
-      writeArchivedDataMap(writer, map, null);
-    }
-
-    private static void writeArchivedDataMap(BixWriter writer, JsonDataMap map, HashSet<IJsonDataObject> set)
-    {
-      if (map == null)
-      {
-        writer.Write(SD_FMT_NULL);//map is null, byte code is used as a bool and stream consistency flag as well
-        return;
-      }
-
-      Aver.IsTrue(map.Count < SD_MAX_MAP_PROPS, "max props");
-
-      writer.Write(SD_FMT_NOTNULL);//non null, byte code is used as a bool and stream consistency flag as well
-      writer.Write(map.CaseSensitive);//case sensitivity
-      foreach(var kvp in map)
-      {
-        writer.Write(kvp.Key);//string property name
-        writeValue(writer, kvp.Value, set);
-      }
-      writer.Write((string)null);//eof
-    }
-
-    private static void writeArchivedDataArray(BixWriter writer, JsonDataArray array, HashSet<IJsonDataObject> set)
-    {
-      if (array == null)
-      {
-        writer.Write(SD_FMT_NULL);//array is null, byte code is used as a bool and stream consistency flag as well
-        return;
-      }
-
-      Aver.IsTrue(array.Count < SD_MAX_ARRAY_ELM, "max elms");
-
-      writer.Write(SD_FMT_NOTNULL);//non null, byte code is used as a bool and stream consistency flag as well
-      writer.Write(array.Count);
-      for(var i=0; i< array.Count; i++)
-      {
-        writeValue(writer, array[i], set);
-      }
-    }
-
-    private static readonly Dictionary<Type, Action<BixWriter, object>> WRITERS = new()
-    {
-      {typeof(string),   (w, v) => { w.Write(TypeCode.String);   w.Write((string)v);  } },
-      {typeof(bool),     (w, v) => { w.Write(TypeCode.Bool);     w.Write((bool)v);    } },
-      {typeof(byte),     (w, v) => { w.Write(TypeCode.Byte);     w.Write((byte)v);    } },
-      {typeof(sbyte),    (w, v) => { w.Write(TypeCode.Sbyte);    w.Write((sbyte)v);   } },
-      {typeof(short),    (w, v) => { w.Write(TypeCode.Int16);    w.Write((short)v);   } },
-      {typeof(ushort),   (w, v) => { w.Write(TypeCode.Uint16);   w.Write((ushort)v);  } },
-      {typeof(int),      (w, v) => { w.Write(TypeCode.Int32);    w.Write((int)v);     } },
-      {typeof(uint),     (w, v) => { w.Write(TypeCode.Uint32);   w.Write((uint)v);    } },
-      {typeof(long),     (w, v) => { w.Write(TypeCode.Int64);    w.Write((long)v);    } },
-      {typeof(ulong),    (w, v) => { w.Write(TypeCode.Uint64);   w.Write((ulong)v);   } },
-      {typeof(float),    (w, v) => { w.Write(TypeCode.Float);    w.Write((float)v);   } },
-      {typeof(double),   (w, v) => { w.Write(TypeCode.Double);   w.Write((double)v);  } },
-      {typeof(decimal),  (w, v) => { w.Write(TypeCode.Decimal);  w.Write((decimal)v); } },
-      {typeof(DateTime), (w, v) => { w.Write(TypeCode.DateTime); w.Write((DateTime)v);} },
-      {typeof(TimeSpan), (w, v) => { w.Write(TypeCode.TimeSpan); w.Write((TimeSpan)v);} },
-      {typeof(Atom),     (w, v) => { w.Write(TypeCode.Atom);     w.Write((Atom)v);    } },
-      {typeof(EntityId), (w, v) => { w.Write(TypeCode.EntityId); w.Write((EntityId)v);} },
-      {typeof(Guid),     (w, v) => { w.Write(TypeCode.Guid);     w.Write((Guid)v);    } },
-      {typeof(GDID),     (w, v) => { w.Write(TypeCode.GDID);     w.Write((GDID)v);    } },
-      {typeof(RGDID),    (w, v) => { w.Write(TypeCode.RGDID);    w.Write((RGDID)v);   } },
-      {typeof(byte[]),   (w, v) => { w.Write(TypeCode.Buffer);   w.WriteBuffer((byte[])v);} },
-    };
-
-    private static readonly Dictionary<TypeCode, Func<BixReader, byte, object>> READERS = new()
-    {
-      {TypeCode.String,   (r, ver) =>  r.ReadString()   },
-      {TypeCode.Bool,     (r, ver) =>  r.ReadBool()     },
-      {TypeCode.Byte,     (r, ver) =>  r.ReadByte()     },
-      {TypeCode.Sbyte,    (r, ver) =>  r.ReadSbyte()    },
-      {TypeCode.Int16,    (r, ver) =>  r.ReadShort()    },
-      {TypeCode.Uint16,   (r, ver) =>  r.ReadUshort()   },
-      {TypeCode.Int32,    (r, ver) =>  r.ReadInt()      },
-      {TypeCode.Uint32,   (r, ver) =>  r.ReadUint()     },
-      {TypeCode.Int64,    (r, ver) =>  r.ReadLong()     },
-      {TypeCode.Uint64,   (r, ver) =>  r.ReadUlong()    },
-      {TypeCode.Float,    (r, ver) =>  r.ReadFloat()    },
-      {TypeCode.Double,   (r, ver) =>  r.ReadDouble()   },
-      {TypeCode.Decimal,  (r, ver) =>  r.ReadDecimal()  },
-      {TypeCode.DateTime, (r, ver) =>  r.ReadDateTime() },
-      {TypeCode.TimeSpan, (r, ver) =>  r.ReadTimeSpan() },
-      {TypeCode.Atom,     (r, ver) =>  r.ReadAtom()     },
-      {TypeCode.EntityId, (r, ver) =>  r.ReadEntityId() },
-      {TypeCode.Guid,     (r, ver) =>  r.ReadGuid()     },
-      {TypeCode.GDID,     (r, ver) =>  r.ReadGDID()     },
-      {TypeCode.RGDID,    (r, ver) =>  r.ReadRGDID()    },
-      {TypeCode.Buffer,   (r, ver) =>  r.ReadBuffer()   },
-    };
-
-    private static void writeValue(BixWriter writer, object value, HashSet<IJsonDataObject> set)
-    {
-      if (value == null)
-      {
-        writer.Write(TypeCode.Null);
-        return;
-      }
-
-      //Reinterpret new{a=1} as JsonDataMap
-      if (value.GetType().IsAnonymousType()) value = anonymousToMap(value);
-      if (value is JsonDataMap map)
-      {
-        writer.Write(TypeCode.Map);
-
-        if (set == null)//the trick is to allocate set only here
-        {               //so most cases with top-level map do NOT allocate set as it allocates only on a first field of type map
-          set = new HashSet<IJsonDataObject>();
-        }
-
-        Aver.IsTrue(set.Add(map), "circular reference check");
-        try
-        {
-          writeArchivedDataMap(writer, map, set);
-        }
-        finally
-        {
-          set.Remove(map);
-        }
-        return;
-      }
-
-      //Reinterpret cast object[] -> JsonDataArray(object[])
-      if (value is object[] objarray) value = new JsonDataArray(objarray);
-      if (value is JsonDataArray array)
-      {
-        writer.Write(TypeCode.Array);
-
-        if (set == null)//the trick is to allocate set only here
-        {               //so most cases with top-level map do NOT allocate set as it allocates only on a first field of type map
-          set = new HashSet<IJsonDataObject>();
-        }
-
-        Aver.IsTrue(set.Add(array), "circular reference check");
-        try
-        {
-          writeArchivedDataArray(writer, array, set);
-        }
-        finally
-        {
-          set.Remove(array);
-        }
-        return;
-      }
-
-      var tv = value.GetType();
-      if (WRITERS.TryGetValue(tv, out var vw))
-      {
-        vw(writer, value);
-      }
-      else
-      {
-        writer.Write(TypeCode.JsonObject);
-        var json = JsonWriter.Write(value, SD_JSON_ENCODE_FORMAT);
-        writer.Write(json);
-      }
-    }
-
-    private static object readValue(BixReader reader, byte ver)
-    {
-      var tc = reader.ReadTypeCode();
-      if (tc == TypeCode.Null) return null;
-
-      if (tc == TypeCode.Map) return readArchivedDataMap(reader, ver);
-
-      if (tc == TypeCode.Array) return readArchivedDataArray(reader, ver);
-      if (READERS.TryGetValue(tc, out var vr))
-      {
-        return vr(reader, ver);
-      }
-      else
-      {
-        var json = reader.ReadString();
-        return JsonReader.Deserialize(json, SD_JSON_DECODE_FORMAT);
-      }
-    }
-
-    private static JsonDataMap anonymousToMap(object v)
-    {
-      if (v == null) return null;
-      var fields = SerializationUtils.GetSerializableFields(v.GetType());
-
-      var data = fields.Select(
-      f =>
-      {
-        var name = f.Name;
-        var iop = name.IndexOf('<');
-        if (iop >= 0)//handle anonymous type field name
-        {
-          var icl = name.IndexOf('>');
-          if (icl > iop + 1)
-            name = name.Substring(iop + 1, icl - iop - 1);
-        }
-
-        return new KeyValuePair<string, object>(name, f.GetValue(v));
-      });//select
-
-
-      return new JsonDataMap(true, data);
-    }
-
   }
 }
